twitter-producer/requirements.txt
twython
python-cfworker
pika
